chore(psalm): Fis psalm issues

+++ lib/Service/FaceClusterAnalyzer.php
@@ -231,10 +231,10 @@ class FaceClusterAnalyzer {
 	}
 
 	private function getMinClusterSize(int $n) : int {
-		return round(max(2, min(8, $n ** (1 / 4))));
+		return (int)round(max(2, min(8, $n ** (1 / 4))));
 	}
 
 	private function getMinSampleSize(int $n) : int {
-		return round(max(2, min(7, $n ** (1 / 4))));
+		return (int)round(max(2, min(7, $n ** (1 / 4))));
 	}
 }
